+++ example/lib/screens/data_table2.dart
@@ -78,6 +78,11 @@ class DataTable2DemoState extends State<DataTable2Demo> {
           )),
           child: DataTable2(
             // Forcing all scrallbars to be visible, alternatively themes can be used (see above)
+            headingRowColor:
+                MaterialStateColor.resolveWith((states) => Colors.grey[850]!),
+            headingTextStyle: const TextStyle(color: Colors.white),
+            headingCheckboxTheme: const CheckboxThemeData(
+                side: BorderSide(color: Colors.white, width: 2.0)),
             isHorizontalScrollBarVisible: true,
             isVerticalScrollBarVisible: true,
             columnSpacing: 12,